cpucount: handle present vs online cpus
Patch from Mark Grondona, with DIAG_ONLINE_CPUCOUNT renamed
to DIAG_CPUCOUNT_ONLINE for naming consistency.

[Mark's comments]
It may be useful to differentiate between cpus that are "present"
versus those that are "online" in the cpucount test. This patch adds
a new DIAG_ONLINE_CPUCOUNT variable which can be set to the expected
number of online cpus in a node. In this case DIAG_CPUCOUNT is compared
against the total number of present cpus while the online cpu count is
compared to DIAG_ONLINE_CPUCOUNT.  If DIAG_ONLINE_CPUCOUNT is not set,
then both the number of present and online cpus are compared to
DIAG_CPUCOUNT (i.e., we assume that online cpus == present cpus).

DIAG_ONLINE_CPUCOUNT may also be set without DIAG_CPUCOUNT, in which
case only the online cpus are counted.

Example (SB node with HT enabled in BIOS, but all HT cpus offline):

 $ grep CPUCOUNT /etc/sysconfig/nodediag
 DIAG_CPUCOUNT="32"
 DIAG_ONLINE_CPUCOUNT="16"

 $ nodediag -v cpucount
 Checking cpucount:                                          [  OK  ]
 1..2
 ok 1 - cpucount: cpucount present 32
 ok 2 - cpucount: cpucount online  16